Group tasks by old or new api endpoints
