+++ src/main/java/org/hisp/dhis/tasks/tracker/AddTrackerEventsTask.java
@@ -1,19 +1,12 @@
-package org.hisp.dhis.tasks.tracker.importer;
+package org.hisp.dhis.tasks.tracker;
 
-import org.hisp.dhis.actions.AuthenticatedApiActions;
-import org.hisp.dhis.actions.RestApiActions;
 import org.hisp.dhis.cache.UserCredentials;
 import org.hisp.dhis.models.Events;
 import org.hisp.dhis.random.EventRandomizer;
-import org.hisp.dhis.random.RandomizerContext;
 import org.hisp.dhis.response.dto.TrackerApiResponse;
 import org.hisp.dhis.tasks.DhisAbstractTask;
-import org.hisp.dhis.tracker.domain.Event;
-import org.hisp.dhis.tracker.domain.mapper.EventMapperImpl;
 import org.hisp.dhis.utils.Randomizer;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.logging.Logger;
 
 /**
