+++ src/main/java/org/hisp/dhis/tasks/tracker/AddTrackerTeiTask.java
@@ -1,9 +1,7 @@
-package org.hisp.dhis.tasks.tracker.importer;
+package org.hisp.dhis.tasks.tracker;
 
-import org.hisp.dhis.cache.User;
 import org.hisp.dhis.cache.UserCredentials;
 import org.hisp.dhis.models.TrackedEntities;
-import org.hisp.dhis.random.RandomizerContext;
 import org.hisp.dhis.response.dto.TrackerApiResponse;
 import org.hisp.dhis.tasks.DhisAbstractTask;
 import org.hisp.dhis.utils.Randomizer;
