+++ src/main/java/org/hisp/dhis/tasks/tracker/oldapi/GenerateAndReserveTrackedEntityAttributeValuesTask.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.tasks.tracker;
+package org.hisp.dhis.tasks.tracker.oldapi;
 
 import org.hisp.dhis.actions.AuthenticatedApiActions;
 import org.hisp.dhis.cache.UserCredentials;
