+++ src/main/java/org/hisp/dhis/tasks/tracker/oldapi/enrollments/AddEnrollmentTask.java
@@ -1,11 +1,8 @@
-package org.hisp.dhis.tasks.tracker.enrollments;
+package org.hisp.dhis.tasks.tracker.oldapi.enrollments;
 
 import org.hisp.dhis.actions.AuthenticatedApiActions;
 import org.hisp.dhis.cache.UserCredentials;
-import org.hisp.dhis.dxf2.events.enrollment.Enrollment;
 import org.hisp.dhis.dxf2.events.enrollment.Enrollments;
-import org.hisp.dhis.random.EnrollmentRandomizer;
-import org.hisp.dhis.random.RandomizerContext;
 import org.hisp.dhis.response.dto.ApiResponse;
 import org.hisp.dhis.tasks.DhisAbstractTask;
 import org.hisp.dhis.utils.Randomizer;
