+++ src/main/java/org/hisp/dhis/tasks/tracker/oldapi/events/AddEventsTask.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.tasks.tracker.events;
+package org.hisp.dhis.tasks.tracker.oldapi.events;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.hisp.dhis.actions.AuthenticatedApiActions;
@@ -6,12 +6,10 @@ import org.hisp.dhis.cache.UserCredentials;
 import org.hisp.dhis.dxf2.events.event.Event;
 import org.hisp.dhis.dxf2.events.event.Events;
 import org.hisp.dhis.random.EventRandomizer;
-import org.hisp.dhis.random.RandomizerContext;
 import org.hisp.dhis.response.dto.ApiResponse;
 import org.hisp.dhis.tasks.DhisAbstractTask;
 import org.hisp.dhis.utils.Randomizer;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Logger;
