+++ src/main/java/org/hisp/dhis/tasks/tracker/oldapi/tei/QueryFilterTeiTask.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.tasks.tracker.tei.oldapi;
+package org.hisp.dhis.tasks.tracker.oldapi.tei;
 
 /*
  * Copyright (c) 2004-2019, University of Oslo
