+++ src/main/java/org/hisp/dhis/tasksets/tracker/Android_syncTeisTaskSet.java
@@ -12,7 +12,7 @@ import org.hisp.dhis.random.RandomizerContext;
 import org.hisp.dhis.random.TrackedEntityInstanceRandomizer;
 import org.hisp.dhis.request.QueryParamsBuilder;
 import org.hisp.dhis.response.dto.ApiResponse;
-import org.hisp.dhis.tasks.tracker.GenerateAndReserveTrackedEntityAttributeValuesTask;
+import org.hisp.dhis.tasks.tracker.oldapi.GenerateAndReserveTrackedEntityAttributeValuesTask;
 import org.hisp.dhis.tasksets.DhisAbstractTaskSet;
 import org.hisp.dhis.utils.Randomizer;
