+++ src/main/java/org/hisp/dhis/tasksets/tracker/importer/Android_importer_syncTeisTaskSet.java
@@ -13,8 +13,8 @@ import org.hisp.dhis.models.TrackerPayload;
 import org.hisp.dhis.random.RandomizerContext;
 import org.hisp.dhis.random.RelationshipRandomizer;
 import org.hisp.dhis.random.TrackedEntityInstanceRandomizer;
-import org.hisp.dhis.tasks.tracker.GenerateAndReserveTrackedEntityAttributeValuesTask;
-import org.hisp.dhis.tasks.tracker.importer.AddTrackerDataTask;
+import org.hisp.dhis.tasks.tracker.oldapi.GenerateAndReserveTrackedEntityAttributeValuesTask;
+import org.hisp.dhis.tasks.tracker.AddTrackerDataTask;
 import org.hisp.dhis.tasksets.DhisAbstractTaskSet;
 import org.hisp.dhis.tracker.domain.TrackedEntity;
 import org.hisp.dhis.tracker.domain.mapper.RelationshipMapperImpl;
