+++ src/main/java/org/hisp/dhis/tasksets/tracker/importer/Capture_importer_addEventTaskSet.java
@@ -6,8 +6,8 @@ import org.hisp.dhis.cache.User;
 import org.hisp.dhis.models.Events;
 import org.hisp.dhis.random.EventRandomizer;
 import org.hisp.dhis.random.RandomizerContext;
-import org.hisp.dhis.tasks.tracker.importer.AddTrackerDataTask;
-import org.hisp.dhis.tasks.tracker.importer.QueryTrackerEventsTask;
+import org.hisp.dhis.tasks.tracker.AddTrackerDataTask;
+import org.hisp.dhis.tasks.tracker.QueryTrackerEventsTask;
 import org.hisp.dhis.tasksets.DhisAbstractTaskSet;
 import org.hisp.dhis.tracker.domain.Event;
 import org.hisp.dhis.tracker.domain.mapper.EventMapperImpl;
