+++ src/main/java/org/hisp/dhis/tasksets/tracker/importer/TrackerCapture_importer_addTeiTaskSet.java
@@ -17,12 +17,12 @@ import org.hisp.dhis.random.TrackedEntityInstanceRandomizer;
 import org.hisp.dhis.response.dto.ApiResponse;
 import org.hisp.dhis.response.dto.TrackerApiResponse;
 import org.hisp.dhis.tasks.DhisDelayedTaskSet;
-import org.hisp.dhis.tasks.tracker.GenerateTrackedEntityAttributeValueTask;
-import org.hisp.dhis.tasks.tracker.importer.AddTrackerDataValueTask;
-import org.hisp.dhis.tasks.tracker.importer.AddTrackerEnrollmentTask;
-import org.hisp.dhis.tasks.tracker.importer.AddTrackerEventsTask;
-import org.hisp.dhis.tasks.tracker.importer.AddTrackerTeiTask;
-import org.hisp.dhis.tasks.tracker.importer.QueryTrackerTeisTask;
+import org.hisp.dhis.tasks.tracker.oldapi.GenerateTrackedEntityAttributeValueTask;
+import org.hisp.dhis.tasks.tracker.AddTrackerDataValueTask;
+import org.hisp.dhis.tasks.tracker.AddTrackerEnrollmentTask;
+import org.hisp.dhis.tasks.tracker.AddTrackerEventsTask;
+import org.hisp.dhis.tasks.tracker.AddTrackerTeiTask;
+import org.hisp.dhis.tasks.tracker.QueryTrackerTeisTask;
 import org.hisp.dhis.tasksets.DhisAbstractTaskSet;
 import org.hisp.dhis.tracker.domain.Attribute;
 import org.hisp.dhis.tracker.domain.Enrollment;
